requirements.txt
beautifulsoup4==4.12.3
matplotlib==3.6.2
numpy==1.24.1
Pillow==10.2.0
requests==2.31.0
tqdm==4.66.1
